% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEM-package.R
\docType{package}
\name{CSEM-package}
\alias{CSEM}
\alias{CSEM-package}
\title{CSEM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Cesar Merino-Soto \email{sikayax@yahoo.com.ar}

Authors:
\itemize{
  \item Cesar Merino-Soto \email{sikayax@yahoo.com.ar}
}

}
\keyword{internal}
