% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvesCSEMcomp.R
\name{print.csem_compare}
\alias{print.csem_compare}
\title{Print method for csem_compare objects}
\usage{
\method{print}{csem_compare}(x, details = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{csem_compare}.}

\item{details}{Logical; if \code{TRUE}, show regression coefficients.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print method for csem_compare objects
}
