% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_data.R
\docType{data}
\name{rse_data}
\alias{rse_data}
\title{Rosenberg Self-Esteem Scale Example Dataset}
\format{
A data frame with 800 rows and 10 variables.
}
\source{
Adapted from the archived package EFA.dimensions v0.1.8.8.
}
\usage{
rse_data
}
\description{
A data frame with 800 observations on 10 items from the
Rosenberg Self-Esteem Scale. Negatively worded items were
reverse scored.
}
\examples{
head(rse_data)

}
\references{
Rosenberg, M. (1965).
\emph{Society and the Adolescent Self-Image}.
Princeton University Press.
}
\keyword{datasets}
