% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCSEM.R
\name{scoreCSEM}
\alias{scoreCSEM}
\title{Obtain CSEM for individual scores from a reference table}
\usage{
scoreCSEM(score.indiv, score.ref, csem, ci = FALSE, conf.level = 0.95)
}
\arguments{
\item{score.indiv}{Numeric vector of individual scores.}

\item{score.ref}{Numeric vector of reference scores (e.g., from a CSEM table).}

\item{csem}{Numeric vector of CSEM values corresponding to \code{score.ref}.}

\item{ci}{Logical. If \code{TRUE}, compute confidence intervals. Default \code{FALSE}.}

\item{conf.level}{Numeric. Confidence level (default 0.95).}
}
\value{
A data frame with the same number of rows as \code{length(score.indiv)},
containing columns: \code{score.indiv}, \code{CSEM}, and if \code{ci = TRUE}, \code{lwr} and \code{upr}.
}
\description{
Given vectors of reference scores and their CSEM values, this function
matches each individual score exactly to a reference score. If an individual
score is not found in the reference, a warning is issued and NA is returned.
Optionally, it computes confidence intervals assuming normality.
}
\examples{
\donttest{
# Use data
library(psychTools)
data(ability)
data.ability <- ability[complete.cases(ability),]

# get CSEM deom Binomial model
res <- csemStrong(data.ability, score.type = "dich", nitems = 16)

# Get CSEM for scores 5, 8, 12 from a strong CSEM table
scoreCsem(score.indiv = c(5, 8, 12),
          score.ref = res$CSEM$raw.score,
          csem = res$CSEM$csem.strong,
          ci = TRUE)
}

}
