% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEM-package.R
\docType{package}
\name{csemTools-package}
\alias{csemTools}
\alias{csemTools-package}
\title{csemTools: Conditional Standard Error of Measurement Tools for Test Scores}
\description{
Provides functions to compute and compare conditional standard errors of measurement (CSEM) across score distributions using multiple statistical approaches from test classical theory. Includes methods for smoothing, CSEM using bootstrapping, standardized CSEM, CSEM for scale score, check propierties of two-halves scores, and comparison with global standard errors of measurement derived from reliability coefficients. The package supports graphical visualization of CSEM curves and relative precision across observed score ranges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmerinos/CSEM}
  \item Report bugs at \url{https://github.com/cmerino/CSEM/issues}
}

}
\author{
\strong{Maintainer}: Cesar Merino-Soto \email{sikayax@yahoo.com.ar}

Authors:
\itemize{
  \item Cesar Merino-Soto \email{sikayax@yahoo.com.ar}
}

}
\keyword{internal}
